moved threads to a seperate component
