--- src/components/testing/serverComponent.tsx
@@ -1,5 +1,0 @@
-'use server' 
-
-export async function GetTime() {
-  return Date.now()
-}
