--- src/components/testing/texts?_locale=en&_quantity=1&_characters=200
@@ -1,1 +1,0 @@
-{"status":"OK","code":200,"locale":"en_US","seed":null,"total":1,"data":[{"title":"So they sat down.","author":"Natasha O'Hara","genre":"Qui","content":"Alice, and she told her sister, as well look and see that the hedgehog had unrolled itself, and was coming back to the door. 'Call the next witness would be of any good reason, and as Alice could."}]}
